add a test for StreamingSoundData::duration
