+++ crates/kira/src/sound/streaming/decoder.rs
@@ -1,3 +1,4 @@
+pub(crate) mod mock;
 #[cfg(feature = "symphonia")]
 pub(crate) mod symphonia;
