+++ crates/kira/src/sound/streaming/sound/test.rs
@@ -8,7 +8,7 @@ use crate::{
 	dsp::Frame,
 	modulator::value_provider::MockModulatorValueProviderBuilder,
 	sound::{
-		streaming::{decoder::Decoder, StreamingSoundData, StreamingSoundSettings},
+		streaming::{decoder::mock::MockDecoder, StreamingSoundData, StreamingSoundSettings},
 		PlaybackState, Sound,
 	},
 	tween::Tween,
@@ -16,56 +16,6 @@ use crate::{
 };
 
 use super::{decode_scheduler::NextStep, StreamingSound};
-
-const MOCK_DECODER_SAMPLE_RATE: u32 = 1;
-const MOCK_DECODER_PACKET_SIZE: usize = 3;
-
-struct MockDecoder {
-	frames: Vec<Frame>,
-	current_frame_index: usize,
-}
-
-impl MockDecoder {
-	fn new(frames: Vec<Frame>) -> Self {
-		Self {
-			frames,
-			current_frame_index: 0,
-		}
-	}
-}
-
-impl Decoder for MockDecoder {
-	type Error = ();
-
-	fn sample_rate(&self) -> u32 {
-		MOCK_DECODER_SAMPLE_RATE
-	}
-
-	fn num_frames(&self) -> usize {
-		self.frames.len()
-	}
-
-	fn decode(&mut self) -> Result<Vec<Frame>, Self::Error> {
-		let mut frames = vec![];
-		for _ in 0..MOCK_DECODER_PACKET_SIZE {
-			frames.push(self.frames[self.current_frame_index]);
-			self.current_frame_index += 1;
-			if self.current_frame_index >= self.frames.len() {
-				break;
-			}
-		}
-		Ok(frames)
-	}
-
-	fn seek(&mut self, index: usize) -> Result<usize, Self::Error> {
-		// seek to the beginning of the "packet" to simulate
-		// seeking behavior with real decoders
-		let index =
-			(index as f64 / MOCK_DECODER_PACKET_SIZE as f64) as usize * MOCK_DECODER_PACKET_SIZE;
-		self.current_frame_index = index;
-		Ok(index)
-	}
-}
 
 /// Tests that a `StreamingSound` will play all of its samples before finishing.
 #[test]
